surveys/serializers.py
from rest_framework import serializers


class LoginSerializer(serializers.Serializer):
    login = serializers.CharField(required=True)
    password = serializers.CharField(required=True, min_length=8)
